requirements.txt
azure-storage-blob>=12.0.0
cloudinary>=1.44.0
google-cloud-vision>=3.0.0
mysql-connector-python>=8.0.0
Pillow>=10.0.0
python-dotenv>=1.0.0
requests>=2.28.0
numpy>=1.24.0
tqdm>=4.65.0
